+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix: Update the confirmation status of send-only (no change) transactions
+
 ## 0.1.0 - 2020-05-20
 
 First release!