+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
-- Fix: Update the confirmation status of send-only (no change) transactions
+- HTTP: Implement the `synced-tip` option to catch up with missed events (#6)
 
 - Unite the `History` event into `Txo{Funded,Spent}`
+
+- Fix: Update the confirmation status of send-only (no change) transactions
 
 ## 0.1.0 - 2020-05-20